feat:add password modification form and modify login check and fix errors
